bugfix: cleanup after resolving merge conflict
