+++ Backend/App/urls.py
@@ -1,13 +1,15 @@
 from django.urls import path
-from .views import wastereportviews, materialtypeviews
+from .views import wastereportviews, materialtypeviews, buildingViews
 
 urlpatterns = [
-    path('materialtype/all', views.get_materialtypes, name='get_materialtypes'),
-    path('wastereport/create', views.post_wastereport, name='post_wastereport'),
-    path('wastereport/all', views.get_all_wastereports, name='get_all_wastereports'),
-    path('wastereport/<int:id>', views.get_wastereport, name='get_wastereport'),
-    path('bygning/byMaterial<str:material>', views.get_allByngingByMaterial, name='get_allByngingByMaterial'),
-    path('bygning/byBygningsnr<int:bygnignsnr>', views.get_singleBygningById, name='get_singleBygningById'),
+    path('wastereport/create', wastereportviews.post_wastereport, name='post_wastereport'),
+    path('wastereport/all', wastereportviews.get_all_wastereports, name='get_all_wastereports'),
+    path('wastereport/<int:id>', wastereportviews.get_wastereport, name='get_wastereport'),
+
+    path('bygning/byMaterial<str:material>', buildingViews.get_allByngingByMaterial, name='get_allByngingByMaterial'),
+    path('bygning/byBygningsnr<int:bygnignsnr>', buildingViews.get_singleBygningById, name='get_singleBygningById'),
+
+    path('materialtype/all', materialtypeviews.get_materialtypes, name='get_materialtypes'),
     path('materialtype/create', materialtypeviews.create_materialtype, name='create_materialtype'),
     path('materialtype/update', materialtypeviews.update_materialtype, name='update_materialtype'),
     path('materialtype/delete', materialtypeviews.delete_materialtype, name='delete_materialtype'),
